feat : 앨범 생성 페이지 추가
앨범 생성 페이지 추가 및 package 위치 변경
