+++ src/main/java/com/bitharmony/comma/domain/album/album/dto/AlbumCreateRequest.java
@@ -1,6 +1,6 @@
-package com.bitharmony.album.dto;
+package com.bitharmony.comma.domain.album.album.dto;
 
-import com.bitharmony.album.entity.Album;
+import com.bitharmony.comma.domain.album.album.entity.Album;
 
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.Size;
