+++ src/main/java/com/bitharmony/comma/domain/album/album/entity/StreamingCount.java
@@ -1,5 +1,8 @@
-package com.bitharmony.album.entity;
+package com.bitharmony.comma.domain.album.album.entity;
 
+import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -13,13 +16,13 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class AlbumLike {
+public class StreamingCount {
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
-	// @JoinColumn(name = "album_id")
-	// private Long albumId;
-	//
-	// @JoinColumn(name = "member_id")
-	// private Long memberId;
+	@Column(nullable = false)
+	private Long streamingDailyCount;
+
+	@Column(name = "streaming_date", nullable = false)
+	private LocalDateTime streamingDate;
 }
