+++ src/main/java/com/bitharmony/comma/domain/album/album/repository/AlbumRepository.java
@@ -1,8 +1,8 @@
-package com.bitharmony.album.repository;
+package com.bitharmony.comma.domain.album.album.repository;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import com.bitharmony.album.entity.Album;
+import com.bitharmony.comma.domain.album.album.entity.Album;
 
 public interface AlbumRepository extends JpaRepository<Album, Long> {
 	Optional<Album> findByAlbumname(String albumname);
