+++ src/main/java/com/bitharmony/comma/domain/album/album/util/NcpConfig.java
@@ -1,4 +1,4 @@
-package com.bitharmony.album.util;
+package com.bitharmony.comma.domain.album.album.util;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
