+++ src/main/resources/application-custom.yml
@@ -1,3 +1,15 @@
 custom:
-  dev-mysql-username:
-  dev-mysql-password:
+  dev-mysql-username: root
+  dev-mysql-password: lldj123414
+cloud:
+  aws:
+    credentials:
+      access-key: q6YAPX1p8y3N8UP6DTCS
+      secret-key: uUbKCAFycuDc6vFinZCnEAOI9y1De6mVTtS4vOHY
+    stack:
+      auto: false
+    region:
+      static: kr-standard
+    s3:
+      endpoint: https://kr.object.ncloudstorage.com
+      bucket: musicstorage
