+++ src/main/resources/application.yml
@@ -1,7 +1,7 @@
 spring:
   profiles:
     active: dev
-    include: secret
+    include: secret, custom
   servlet:
     multipart:
       max-file-size: 50MB
