+++ src/main/resources/templates/domain/album/album_form.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Album Form</title>
+    <style>
+        #licenseDescription, #licenseDescriptionLabel, #price, #priceLabel {
+            display: none;
+        }
+    </style>
+</head>
+<body>
+
+<form method="post" enctype="multipart/form-data">
+    <label>Album Name</label>
+    <input type="text" name="albumName"> <br>
+
+    <label>Genre</label>
+    <select name="genre">
+        <option value="rock">Rock</option>
+        <option value="pop">Pop</option>
+        <option value="jazz">Jazz</option>
+        <option value="classical">Classical</option>
+        <!-- Add more options as needed -->
+    </select> <br>
+
+    <label>Music File</label>
+    <input type="file" id="musicFile" name="musicFile"> <br>
+
+    <label>Music Img File</label>
+    <input type="file" id="musicImageFile" name="musicImageFile"> <br>
+
+    <label>License</label>
+    <input type="checkbox" id="license" name="license"> <br>
+
+    <label id="licenseDescriptionLabel">License Description</label>
+    <textarea id="licenseDescription" name="licenseDescription"></textarea> <br>
+
+    <label>Permit</label>
+    <input type="checkbox" id="permit" name="permit"> <br>
+
+    <label id="priceLabel">Price</label>
+    <input type="number" id="price" name="price"> <br>
+    <input type="submit" value="Submit">
+</form>
+
+<script>
+    document.getElementById('license').addEventListener('change', function() {
+        var licenseDescription = document.getElementById('licenseDescription');
+        var licenseDescriptionLabel = document.getElementById('licenseDescriptionLabel');
+        if (this.checked) {
+            licenseDescription.style.display = 'block';
+            licenseDescriptionLabel.style.display = 'block';
+        } else {
+            licenseDescription.style.display = 'none';
+            licenseDescriptionLabel.style.display = 'none';
+        }
+    });
+
+    document.getElementById('permit').addEventListener('change', function() {
+        var price = document.getElementById('price');
+        var priceLabel = document.getElementById('priceLabel');
+        if (this.checked) {
+            price.style.display = 'block';
+            priceLabel.style.display = 'block';
+        } else {
+            price.style.display = 'none';
+            priceLabel.style.display = 'none';
+        }
+    });
+</script>
+
+</body>
+</html>
